Seperate other store to own file
